dedicated executable + setup.py
