+++ git_of_theseus/__init__.py
@@ -1,0 +1,3 @@
+from git_of_theseus.stack_plot import stack_plot
+from git_of_theseus.survival_plot import survival_plot
+from git_of_theseus.analyze import analyze
