+++ scripts/git-of-theseus-analyze
@@ -1,0 +1,3 @@
+#!/usr/bin/python
+from git_of_theseus import analyze
+analyze()
