+++ scripts/git-of-theseus-stack-plot
@@ -1,0 +1,3 @@
+#!/usr/bin/python
+from git_of_theseus import stack_plot
+stack_plot()
