+++ scripts/git-of-theseus-survival-plot
@@ -1,0 +1,3 @@
+#!/usr/bin/python
+from git_of_theseus import survival_plot
+survival_plot()
